break(iot-deps): Remove deps package, move shared code to respective packages
For instance, classes like IotHubSSLContext that were used in both IoT Hub device client and IoT Hub service packages are copied to both packages.

A notable benefit to this approach is that the dependencies we had in the iot-deps package (mqtt and amqp library in particular) will no longer needlessly be brought in by the provisioning service client package which used to have a dependency on the iot-deps package

+++ device/iot-device-client/src/test/java/com/microsoft/azure/sdk/iot/device/edge/MethodRequestTest.java
@@ -5,7 +5,6 @@
 
 package com.microsoft.azure.sdk.iot.device.edge;
 
-import com.microsoft.azure.sdk.iot.device.edge.MethodRequest;
 import mockit.Deencapsulation;
 import org.junit.Test;
 
@@ -57,8 +56,8 @@ public class MethodRequestTest
         MethodRequest request = new MethodRequest(expectedMethodName, expectedPayload, expectedResponseTimeout, expectedConnectionTimeout);
 
         //assert
-        assertEquals(expectedResponseTimeout, Deencapsulation.getField(request, "responseTimeoutInSeconds"));
-        assertEquals(expectedConnectionTimeout, Deencapsulation.getField(request, "connectionTimeoutInSeconds"));
+        assertEquals(expectedResponseTimeout, (int) Deencapsulation.getField(request, "responseTimeoutInSeconds"));
+        assertEquals(expectedConnectionTimeout, (int) Deencapsulation.getField(request, "connectionTimeoutInSeconds"));
         assertEquals(expectedMethodName, Deencapsulation.getField(request, "methodName"));
         assertEquals(expectedPayload, Deencapsulation.getField(request, "payload"));
